chore: spring 모듈 제거
CommandHandlers는 스프링을 사용하는 이상 다른 방식으로 구현될 일이 없기 때문에 api 모듈의 것이 맞다고 판단
기존에 구현된 CommandBus(InMemory)는 단일 서버인 main-api가 기본으로 사용하기 때문에 api에 있는 것이 맞다고 판단
