--- modules/infrastructure/spring/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'com.Whoz-In'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compileOnly project(':modules:domain')
-    compileOnly project(':modules:main-api')
-
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-test {
-    useJUnitPlatform()
-}
