+++ modules/main-api/build.gradle
@@ -28,7 +28,6 @@ dependencies {
     implementation project(":modules:infrastructure:logging")
 
     //main-api 구현 모듈 (runtimeOnly - 순환 의존 방지)
-    runtimeOnly project(":modules:infrastructure:spring")
     runtimeOnly project(":modules:infrastructure:api-query-jpa")
 
     testImplementation platform('org.junit:junit-bom:5.9.1')
