+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/application/command/InMemoryCommandBus.java
@@ -1,0 +1,17 @@
+package com.whoz_in.main_api.shared.application.command;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public final class InMemoryCommandBus implements CommandBus {
+    private final CommandHandlers handlers;
+
+    @Override
+    public void dispatch(Command command){
+        CommandHandler handler = handlers.findBy(command);
+        handler.handle(command);
+    }
+}
+
