+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/application/query/QueryHandlers.java
@@ -1,7 +1,5 @@
-package com.whoz_in.spring.application.query;
+package com.whoz_in.main_api.shared.application.query;
 
-import com.whoz_in.main_api.shared.application.query.Query;
-import com.whoz_in.main_api.shared.application.query.QueryHandler;
 import java.lang.reflect.ParameterizedType;
 import java.util.Map;
 import java.util.stream.Collectors;
